+++ gitutils/const.py
@@ -1,6 +1,16 @@
 from urllib.parse import urljoin
 
 
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
 ######################
 # NOTIFICATION LEVEL #
 ######################
@@ -23,6 +33,9 @@ ENDPOINT = "https://git.psi.ch"
 OAUTH_ROUTE = urljoin(ENDPOINT, "/oauth/token")
 OATH_REQUEST = urljoin(OAUTH_ROUTE, "?grant_type=password&username=")
 PASSWORD_URL = "&password="
+
+SEARCH_ROUTE= urljoin(ENDPOINT, "/search?")
+CODE_URL ="&search_code=true"
 ############
 # MESSAGES #
 ############
@@ -50,6 +63,7 @@ FORK_PROBLEM_REMOTE = "Existing repo with the same name. Please indicate -c if y
 FORK_PROBLEM_GIT_FOLDER = "Fork is not recommended inside a repository folder. Please execute the fork command outside a repository folder."
 FORK_PROBLEM_PERSONAL = "Not possible to fork your own personal project."
 CLONEGROUP_PROBLEM = "Not possible to clonegroup if the group name parameter is not provided. Use 'gitutils clonegroup -h' for help."
+SEARCHFILE_PROBLEM = "Not possible to search for a file if the group or file are not provided. Use 'gitutils search -h' for help."
 FORK_GROUP_NOT_FOUND = "The requested fork project could not be found under the specified group or check if you have permission to access such group/project."
 GROUP_NAME_PROBLEM = "The requested project was found under a different group or check if you have permission to access such group/project."
 PROJECT_URL_NOT_FOUND = "Project url not found, please check the configuration details or check if you have permission to access such group/project.."
@@ -121,7 +135,19 @@ MERGE_MESSAGE_TITLE = ''' The title of the merge request that is going to be cre
 MERGE_MESSAGE_DESCRIPTION = '''The description of the merge request that is going to be created.'''
 CLONEGROUP_HELP_MSG = "Clones all existing projects within a group."
 CLONEGROUP_GROUP_NAME = "Group name"
+SEARCHFILE_HELP_MSG = "Search for a file within a group."
+SEARCHFILE_FILE_MSG = "File's name."
+GREP_PROJECT_MSG = "Project's name."
+GREP_TERM_MSG = "Term to search."
+SEARCHFILE_GROUP_MSG = "Group's name"
+SEARCHFILE_INIT_MSG = '\nGitutils searching inside group%s %s %sfor file%s %s %s...'
+GREPFILE_INIT_MSG = '\nGitutils searching inside project%s %s %sfor term \"%s %s %s\"...'
+GREPFILE_PROBLEM = "Not possible to search for a term if the project or the term itself are not provided. Use 'gitutils grep -h' for help."
+GREPFILE_HELP_MSG = "Search for a term inside the files of a project."
+GREP_EMPTY = "The term %s %s %s was not found in such project."
+SEARCHFILE_EMPTY = "The file %s %s %s was not found in such group."
 STORE_TRUE = "store_true"
+BLOBS = 'blobs'
 ############
 # COMMANDS #
 ############